Add WowJournalInstance model and populate it

+++ apps/tool/Models/Journal/DumpJournalInstance.cs
@@ -5,11 +5,11 @@ namespace Wowthing.Tool.Models.Journal;
 
 public class DumpJournalInstance
 {
-    public int  ID { get; set; }
+    public short  ID { get; set; }
 
     public int AreaID { get; set; }
     public int Flags { get; set; }
-    public int MapID { get; set; }
+    public short MapID { get; set; }
 
     [Name("Name_lang")]
     public string Name { get; set; } = string.Empty;
