+++ apps/tool/Models/Journal/DumpJournalTierXInstance.cs
@@ -7,7 +7,7 @@ public class DumpJournalTierXInstance
 {
     public int ID { get; set; }
 
-    public int JournalInstanceID { get; set; }
+    public short JournalInstanceID { get; set; }
     public int JournalTierID { get; set; }
     public int OrderIndex { get; set; }
